% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularitree.R
\name{generate_sample}
\alias{generate_sample}
\title{Generate a sample of species, weighted by their score}
\usage{
generate_sample(taxon_scores, ntax, phy, pd_weighting, min.prob = 0.01,
  max.prob = 0.99)
}
\arguments{
\item{taxon_scores}{The vector of scores for taxa}

\item{ntax}{The number of taxa you want returned}

\item{phy}{Phylogeny to prune}

\item{pd_weighting}{How much to weight the pd vs the popularity}

\item{min.prob}{The minimum probability of sampling for a taxon}

\item{max.prob}{The maximum probability of sampling for a taxon}
}
\value{
A list: selected=vector of names of sampled taxa to include, and score=sum of the taxon metrics for these taxa
}
\description{
Generate a sample of species, weighted by their score
}
