% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularitree.R
\name{popularitree}
\alias{popularitree}
\title{Get a smaller tree with popular taxa spanning the diversity}
\usage{
popularitree(phy, ntax, measures, pd_weighting = 0.5, nrep = 10000)
}
\arguments{
\item{phy}{A phylo object}

\item{ntax}{The final number of taxa you want}

\item{measures}{Data.frame of the measures of popularity}

\item{pd_weighting}{How much to weight the pd vs the popularity}

\item{nrep}{How many replicates to use}
}
\value{
A phylo object pruned to the number of taxa
}
\description{
This assumes you will have a data.frame of popularity measures
and then sum to combine the columns into a single measure.
}
\examples{
big.phy <- rotl::get_study("ot_308")[[1]]
big.phy <- geiger::drop.random(big.phy, ape::Ntip(phy) - 20)
measures <- tree_compute_measures(big.phy)
small.phy <- popularitree(phy=big.phy, ntax=5, measures, nrep=100)
par(mfcol=c(1,3))
plot(big.phy)
plot(small.phy)
plot(tree_convert_to_common(small.phy))
}
