% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularitree.R
\name{species_get_common}
\alias{species_get_common}
\title{Return common name if known, otherwise return scientific name}
\usage{
species_get_common(species)
}
\arguments{
\item{species}{Species name}
}
\value{
Common name (ideally) or species name
}
\description{
Return common name if known, otherwise return scientific name
}
