% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularitree.R
\name{species_gtrends_popularity}
\alias{species_gtrends_popularity}
\title{Get a sum of hits of a species relative to a standard}
\usage{
species_gtrends_popularity(species, standard = "Myrmeocystus mexicanus")
}
\arguments{
\item{species}{The focal species to look at search interest over}

\item{standard}{Since Google Trends always compares popularity, give a standard species to use}
}
\value{
The sum of relative hits over the time period
}
\description{
Get a sum of hits of a species relative to a standard
}
