% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularitree.R
\name{species_ncbi_common_exists}
\alias{species_ncbi_common_exists}
\title{Return 1 if a species has a common name in NCBI, 0 otherwise}
\usage{
species_ncbi_common_exists(species)
}
\arguments{
\item{species}{The focal species to examine}
}
\value{
1 or 0
}
\description{
Return 1 if a species has a common name in NCBI, 0 otherwise
}
\examples{
species_ncbi_common_exists('Puma concolor')
species_ncbi_common_exists('Pomatomus saltatrix')

}
