% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularitree.R
\name{tree_compute_measures}
\alias{tree_compute_measures}
\title{Compute popularity measures}
\usage{
tree_compute_measures(phy)
}
\arguments{
\item{phy}{An input phylogeny}
}
\value{
data.frame of measures
}
\description{
Compute popularity measures
}
