% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popularitree.R
\name{tree_convert_to_common}
\alias{tree_convert_to_common}
\title{Convert tips to common names if possible}
\usage{
tree_convert_to_common(phy)
}
\arguments{
\item{phy}{A phylogeny}
}
\value{
Phylogeny with names changed
}
\description{
Convert tips to common names if possible
}
